+++ railties/test/application/configuration_test.rb
@@ -754,9 +754,7 @@ module ApplicationTests
 
     test "Use key_generator when secret_key_base is set" do
       make_basic_app do |application|
-        Rails.deprecator.silence do
-          application.secrets.secret_key_base = "b3c631c314c0bbca50c1b2843150fe33"
-        end
+        application.config.secret_key_base = "b3c631c314c0bbca50c1b2843150fe33"
         application.config.session_store :disabled
       end
 
@@ -776,9 +774,7 @@ module ApplicationTests
 
     test "application verifier can be used in the entire application" do
       make_basic_app do |application|
-        Rails.deprecator.silence do
-          application.secrets.secret_key_base = "b3c631c314c0bbca50c1b2843150fe33"
-        end
+        application.config.secret_key_base = "b3c631c314c0bbca50c1b2843150fe33"
         application.config.session_store :disabled
       end
 
@@ -921,112 +917,14 @@ module ApplicationTests
       assert_equal "old message", app.message_verifiers["salt"].verify(old_message)
     end
 
-    test "secrets is deprecated" do
-      app "development"
-
-      assert_deprecated(Rails.deprecator) do
-        Rails.application.secrets.secret_key_base = "3b7cd727ee24e8444053437c36cc66c3"
-      end
-    end
-
-    test "secrets.secret_key_base is used when config/secrets.yml is present" do
-      remove_file "config/credentials.yml.enc"
-      app_file "config/secrets.yml", <<-YAML
-        development:
-          secret_key_base: 3b7cd727ee24e8444053437c36cc66c3
-      YAML
-
-      app "development"
-      Rails.deprecator.silence do
-        assert_equal "3b7cd727ee24e8444053437c36cc66c3", app.secrets.secret_key_base
-      end
-      assert_equal "3b7cd727ee24e8444053437c36cc66c3", app.secret_key_base
-    end
 
     test "secret_key_base is copied from config.secret_key_base when set" do
-      remove_file "config/secrets.yml"
       app_file "config/initializers/secret_token.rb", <<-RUBY
         Rails.application.config.secret_key_base = "3b7cd727ee24e8444053437c36cc66c3"
       RUBY
 
       app "development"
       assert_equal "3b7cd727ee24e8444053437c36cc66c3", app.secret_key_base
-    end
-
-    test "config.secret_key_base does not lead to a deprecation" do
-      remove_file "config/secrets.yml"
-      app_file "config/initializers/secret_token.rb", <<-RUBY
-        Rails.application.credentials.secret_key_base = nil
-        Rails.application.config.secret_key_base = "3b7cd727ee24e8444053437c36cc66c3"
-      RUBY
-
-      app "production"
-
-      assert_not_deprecated(Rails.deprecator) do
-        assert_equal "3b7cd727ee24e8444053437c36cc66c3", app.secret_key_base
-      end
-    end
-
-    test "config.secret_key_base leads to a deprecation in development when config/secrets.yml is present" do
-      remove_file "config/credentials.yml.enc"
-      app_file "config/secrets.yml", <<-YAML
-        development:
-          secret_key_base: 3b7cd727ee24e8444053437c36cc66c3
-      YAML
-
-      app "development"
-      assert_deprecated(Rails.deprecator) do
-        assert_equal "3b7cd727ee24e8444053437c36cc66c3", app.secrets.secret_key_base
-      end
-      assert_equal "3b7cd727ee24e8444053437c36cc66c3", app.secret_key_base
-    end
-
-    test "custom secrets saved in config/secrets.yml are loaded in app secrets" do
-      app_file "config/secrets.yml", <<-YAML
-        development:
-          secret_key_base: 3b7cd727ee24e8444053437c36cc66c3
-          aws_access_key_id: myamazonaccesskeyid
-          aws_secret_access_key: myamazonsecretaccesskey
-      YAML
-
-      app "development"
-
-      assert_equal "myamazonaccesskeyid", app.secrets.aws_access_key_id
-      assert_equal "myamazonsecretaccesskey", app.secrets.aws_secret_access_key
-    end
-
-    test "shared secrets saved in config/secrets.yml are loaded in app secrets" do
-      app_file "config/secrets.yml", <<-YAML
-        shared:
-          api_key: 3b7cd727
-      YAML
-
-      app "development"
-
-      assert_equal "3b7cd727", app.secrets.api_key
-    end
-
-    test "shared secrets will yield to environment specific secrets" do
-      app_file "config/secrets.yml", <<-YAML
-        shared:
-          api_key: 3b7cd727
-
-        development:
-          api_key: abc12345
-      YAML
-
-      app "development"
-
-      assert_equal "abc12345", app.secrets.api_key
-    end
-
-    test "blank config/secrets.yml does not crash the loading process" do
-      app_file "config/secrets.yml", <<-YAML
-      YAML
-
-      app "development"
-
-      assert_nil app.secrets.not_defined
     end
 
     test "config.secret_key_base over-writes a blank app.secret_key_base" do
@@ -1037,20 +935,6 @@ module ApplicationTests
       app "development"
 
       assert_equal "iaminallyoursecretkeybase", app.secret_key_base
-    end
-
-    test "that nested keys are symbolized the same as parents for hashes more than one level deep" do
-      app_file "config/secrets.yml", <<-YAML
-        development:
-          smtp_settings:
-            address: "smtp.example.com"
-            user_name: "postmaster@example.com"
-            password: "697361616320736c6f616e2028656c6f7265737429"
-      YAML
-
-      app "development"
-
-      assert_equal "697361616320736c6f616e2028656c6f7265737429", app.secrets.smtp_settings[:password]
     end
 
     test "require_master_key aborts app boot when missing key" do